requirements.txt
matplotlib>=3.0.0
pandas>=1.0.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
fastapi>=0.100.0
uvicorn>=0.22.0
circuitbreaker>=1.4.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0